Added ports and updated hardware code for build fixes

+++ src/com/palyrobotics/frc2017/config/RobotState.java
@@ -36,5 +36,5 @@ public class RobotState {
 
 	// Climber
 	public double climberCurrentDraw;
-	public int climberEncoder;
+	public double climberEncoder;
 }
